fix[mbpp]: 244-300

+++ groundtruth/mbpp/267.py
@@ -5,7 +5,7 @@ Write a python function that takes in an integer n and returns the sum of the sq
 def square_Sum(n):  
     assert isinstance(n, int), "invalid inputs" # $_CONTRACT_$
     assert n > 0, "invalid inputs" # $_CONTRACT_$
-    return int(n*(4*n*n-1)/3) 
+    return n * (4 * n * n - 1) / 3
